Created a weighted call graph and removed redudant code
