+++ app2callgraph
@@ -25,5 +25,4 @@ while [ "$1" != "" ]; do
     shift
 done
 
-
-eval "./gradlew --quiet clean code2graph -Pargs=\"$ARGS\""
+eval "./gradlew --quiet clean app2callgraph -Pargs=\"$ARGS\""
