+++ src/main/java/com/ibm/minerva/app2callgraph/utils/AnalysisUtils.java
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package org.konveyor.dgi.code2graph.utils;
+package com.ibm.minerva.app2callgraph.utils;
 
 import com.ibm.wala.classLoader.IClass;
 import com.ibm.wala.classLoader.IMethod;
@@ -68,30 +68,5 @@ public class AnalysisUtils {
     }
     Log.info("Registered " + entrypointsCount + " entrypoints.");
     return entrypoints;
-  }
-
-  public static void expandSymbolTable(IClassHierarchy cha) {
-    for (IClass c : cha) {
-      if (isApplicationClass(c)) {
-        // private constructor
-        String className = c.getName().getClassName().toString();
-        Map<String, Object> classAttributeMap = new HashMap<String, Object>();
-        String classIsPrivate = Boolean.toString(c.isPrivate());
-        classAttributeMap.put("isPrivate", classIsPrivate);
-        String classSourcePath = c.getSourceFileName();
-        classAttributeMap.put("source_file_name", classSourcePath);
-        Integer num_fields = c.getAllFields().size();
-        classAttributeMap.put("num_fields", num_fields);
-        Integer num_static_fields = c.getAllStaticFields().size();
-        classAttributeMap.put("num_static_fields", num_static_fields);
-        Integer num_static_methods = 0;
-        for (IMethod method : c.getDeclaredMethods()) {
-          num_static_methods += method.isStatic() ? 1 : 0;
-        }
-        classAttributeMap.put("num_static_methods", num_static_methods);
-        classAttributeMap.put("num_declared_methods", c.getDeclaredMethods().size());
-        classAttr.put(className, classAttributeMap);
-      }
-    }
   }
 }
