--- src/main/java/org/konveyor/dgi/code2graph/utils/graph/MethodNode.java
@@ -1,4 +1,0 @@
-package org.konveyor.dgi.code2graph.utils.graph;
-
-public class MethodNode {
-}
